tests/wpt/web-platform-tests/trusted-types/Location-href.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  async_test(t => {
    createURL_policy(window)
      .then(t.step_func_done(p => {
        let url = p.createURL(location.href + "#xxx");
        location.href = url;
        assert_equals("" + url, location.href, "location href");
    }));
  }, "location.href assigned via policy (successful URL transformation).");
</script>
